refatoração
Correções e redesign visual completo (Brasa Burger — hamburgueria).

Backend / correções:
- .gitignore + remoção de bin/obj/Movel.db do versionamento
- migração .NET 7 (EOL) -> .NET 8; remove pacote SqlServer sem uso
- Item.Preco double -> decimal (+ migration PrecoParaDecimal); corrige
  InvalidCastException no total do checkout; Sum() de decimal no cliente
- segurança: antiforgery no Login/Logout, lockout contra brute force,
  [Bind] no Checkout, credenciais do admin via configuração (não hardcoded)
- bugs: link do card (Detail->Details), busca com Contains, tag helpers
  quebrados (data-bs- slide, asp- action, controller CarrinhoCompra),
  <img src> incompleto no CheckoutCompleto, connection string DefaulConnection

Redesign visual (ui-ux-pro-max):
- design system em site.css (paleta quente, Bebas Neue + Poppins, tokens)
- navbar, hero, cards de cardápio, carrinho, checkout, login/registro,
  área admin e footer refeitos; acessibilidade (foco, alt, labels, skip link)

+++ VewModel/CarrinhoViewModel.cs
@@ -6,6 +6,6 @@ namespace PROJETO.VewModel
     public class CarrinhoViewModel 
     {
         public Carrinho Carrinho { get; set; }
-        public double CarrinhoTotal { get; set; }
+        public decimal CarrinhoTotal { get; set; }
     }
 }
